fix race condition introduced by docker 1.3.0
It appears docker 1.3.0 uncovered a race condition that was not
previously seen where there is a small delay between container creation
and container startup when `docker run` is executed that could result in
`docker inspect` returning successfully but before the container is
running. When this occurs there is no networking information such as IP
of the container and port mappings. This is probably a side effect of
the new `docker create` command which makes creation and running
discrete actions now.

The solution is to introduce a `wait_until_running` function which will
run `docker inspect` in a bounded loop until the State of the container
turns from `running = false` into `running = true`, then the util will
proceed to gather the port mappings and setup the iptables rules.